feat(dashboards): Remove assignee selector column
This is the last place we're using the old assignee selector. It overflows the widget. This might not be the best place to assign an issue.

Instead, render just the current assignee.
