+++ static/app/actionCreators/group.tsx
@@ -4,58 +4,13 @@ import type {RequestCallbacks, RequestOptions} from 'sentry/api';
 import {Client} from 'sentry/api';
 import GroupStore from 'sentry/stores/groupStore';
 import type {Actor} from 'sentry/types/core';
-import type {Group, Note, Tag as GroupTag, TagValue} from 'sentry/types/group';
-import type {Member} from 'sentry/types/organization';
-import type {User} from 'sentry/types/user';
+import type {Tag as GroupTag, TagValue} from 'sentry/types/group';
 import {buildTeamId, buildUserId} from 'sentry/utils';
 import {uniqueId} from 'sentry/utils/guid';
 import type {ApiQueryKey, UseApiQueryOptions} from 'sentry/utils/queryClient';
 import {useApiQuery} from 'sentry/utils/queryClient';
 
 type AssignedBy = 'suggested_assignee' | 'assignee_selector';
-type AssignToUserParams = {
-  assignedBy: AssignedBy;
-  /**
-   * Issue id
-   */
-  id: string;
-  orgSlug: string;
-  user: User | Actor;
-  member?: Member;
-};
-
-export function assignToUser(params: AssignToUserParams) {
-  const api = new Client();
-
-  const endpoint = `/organizations/${params.orgSlug}/issues/${params.id}/`;
-
-  const id = uniqueId();
-
-  GroupStore.onAssignTo(id, params.id, {
-    email: params.member?.email ?? '',
-  });
-
-  const request = api.requestPromise(endpoint, {
-    method: 'PUT',
-    // Sending an empty value to assignedTo is the same as "clear",
-    // so if no member exists, that implies that we want to clear the
-    // current assignee.
-    data: {
-      assignedTo: params.user ? buildUserId(params.user.id) : '',
-      assignedBy: params.assignedBy,
-    },
-  });
-
-  request
-    .then(data => {
-      GroupStore.onAssignToSuccess(id, params.id, data);
-    })
-    .catch(data => {
-      GroupStore.onAssignToError(id, params.id, data);
-    });
-
-  return request;
-}
 
 export function clearAssignment(
   groupId: string,
@@ -139,70 +94,6 @@ export function assignToActor({id, actor, assignedBy, orgSlug}: AssignToActorPar
     .catch(data => {
       GroupStore.onAssignToSuccess(guid, id, data);
     });
-}
-
-export function deleteNote(
-  api: Client,
-  orgSlug: string,
-  group: Group,
-  id: string,
-  _oldText: string
-) {
-  const restore = group.activity.find(activity => activity.id === id);
-  const index = GroupStore.removeActivity(group.id, id);
-
-  if (index === -1 || restore === undefined) {
-    // I dunno, the id wasn't found in the GroupStore
-    return Promise.reject(new Error('Group was not found in store'));
-  }
-
-  const promise = api.requestPromise(
-    `/organizations/${orgSlug}/issues/${group.id}/comments/${id}/`,
-    {
-      method: 'DELETE',
-    }
-  );
-
-  promise.catch(() => GroupStore.addActivity(group.id, restore, index));
-
-  return promise;
-}
-
-export function createNote(api: Client, orgSlug: string, group: Group, note: Note) {
-  const promise = api.requestPromise(
-    `/organizations/${orgSlug}/issues/${group.id}/comments/`,
-    {
-      method: 'POST',
-      data: note,
-    }
-  );
-
-  promise.then(data => GroupStore.addActivity(group.id, data));
-
-  return promise;
-}
-
-export function updateNote(
-  api: Client,
-  orgSlug: string,
-  group: Group,
-  note: Note,
-  id: string,
-  oldText: string
-) {
-  GroupStore.updateActivity(group.id, id, {text: note.text});
-
-  const promise = api.requestPromise(
-    `/organizations/${orgSlug}/issues/${group.id}/comments/${id}/`,
-    {
-      method: 'PUT',
-      data: note,
-    }
-  );
-
-  promise.catch(() => GroupStore.updateActivity(group.id, id, {text: oldText}));
-
-  return promise;
 }
 
 type ParamsType = {
